requirements.txt
pandas>=2.0.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
great-expectations>=0.18.0
pytest>=7.0.0
pytest-cov>=4.0.0
pydantic>=2.0.0
structlog>=23.0.0